1. Added Different tabs for morse and pineconedocs 2. Added logic to hit namespaces 3. Open bug layout issues
