+++ .gitignore
@@ -38,3 +38,5 @@ next-env.d.ts
 
 #Notion_db
 /Notion_DB
+
+.history
