+++ package.json
@@ -18,6 +18,7 @@
     "ingest-data-markdown": "tsx -r dotenv/config scripts/ingest-data-markdown.ts"
   },
   "dependencies": {
+    "@chatscope/chat-ui-kit-react": "^1.10.1",
     "@microsoft/fetch-event-source": "^2.0.1",
     "@pinecone-database/pinecone": "0.0.12",
     "@radix-ui/react-accordion": "^1.1.1",
@@ -40,6 +41,7 @@
     "react-markdown": "^8.0.5",
     "sharp": "^0.32.1",
     "tailwind-merge": "^1.10.0",
+    "ts-node": "^10.9.1",
     "turndown": "^7.1.2",
     "uuidv4": "^6.2.13"
   },
@@ -55,7 +57,8 @@
     "prettier": "^2.8.4",
     "tailwindcss": "^3.2.7",
     "tsx": "^3.12.3",
-    "typescript": "^4.9.5"
+    "typescript": "^4.9.5",
+    "typescript-plugin-css-modules": "^5.0.1"
   },
   "engines": {
     "node": ">=18"
