+++ pages/_app.tsx
@@ -1,4 +1,6 @@
 import '@/styles/base.css';
+import '@/styles/globals.css';
+//import '@/styles/tailwind.css';
 import type { AppProps } from 'next/app';
 import { Inter } from 'next/font/google';
