+++ utils/makechain.ts
@@ -8,6 +8,7 @@ Chat History:
 {chat_history}
 Follow Up Input: {question}
 Standalone question:`;
+console.log('🚀 ~ file: makechain.ts:11 ~ CONDENSE_PROMPT:', CONDENSE_PROMPT);
 
 const QA_PROMPT = `You are a helpful AI assistant. Use the following pieces of context to answer the question at the end.
 If you don't know the answer, just say you don't know. DO NOT try to make up an answer.
@@ -20,7 +21,7 @@ Helpful answer in markdown:`;
 
 export const makeChain = (vectorstore: PineconeStore) => {
   const model = new OpenAI({
-    temperature: 0.3, // increase temepreature to get more creative answers
+    temperature: 0, // increase temepreature to get more creative answers
     modelName: 'gpt-3.5-turbo', //change this to gpt-4 if you have access
   });
 
@@ -30,8 +31,9 @@ export const makeChain = (vectorstore: PineconeStore) => {
     {
       qaTemplate: QA_PROMPT,
       questionGeneratorTemplate: CONDENSE_PROMPT,
-      returnSourceDocuments: true, //The number of source documents returned is 4 by default
+      returnSourceDocuments: true,
     },
   );
+
   return chain;
 };
